shared-backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
supabase>=2.4.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
PyJWT>=2.8.0
python-multipart>=0.0.9
httpx>=0.27.0
